+++ lib/acts_as_ditto.rb
@@ -4,13 +4,9 @@ require "active_support"
 
 require_relative "ditto/version"
 require_relative "ditto/configuration"
-require_relative "ditto/duplication_context"
-require_relative "ditto/duplicator"
+require_relative "ditto/duplication/duplication_context"
+require_relative "ditto/duplication/duplicator"
 require_relative "ditto/acts_as_ditto"
-
-module Ditto
-  class Error < StandardError; end
-end
 
 ActiveSupport.on_load(:active_record) do
   include Ditto::ActsAsDitto